requirements/stable-v0.1.20.txt
ambition-ae==0.1.8
ambition-dashboard==0.1.2a5
ambition-export==0.1.3
ambition-labs==0.1.5
ambition-metadata-rules==0.1.5
ambition-pharmacy==0.1.3
ambition-prn==0.1.6
ambition-rando==0.1.10
ambition-reference==0.1.5
ambition-screening==0.1.6
ambition-sites==0.1.5
ambition-subject==0.1.15
ambition-validators==0.1.5
ambition-visit-schedule==0.1.6
edc-action-item==0.1.7
edc-appointment==0.2.18
edc-auth==0.1.8
edc-base==0.2.29
edc-consent==0.2.13
edc-constants==0.2.10
edc-dashboard==0.2.10
edc-device==0.2.8
edc-fabric==0.1.5a1
edc-facility==0.1.8
edc-fieldsets==0.1.3
edc-form-validators==0.1.5
edc-identifier==0.2.14
edc-lab==0.2.24
edc-lab-dashboard==0.1.15
edc-label==0.1.11
edc-list-data==0.1.2
edc-locator==0.2.9
edc-metadata==0.2.16
edc-metadata-rules==0.1.12
edc-model-admin==0.1.6
edc-model-fields==0.1.4
edc-model-wrapper==0.1.9
edc-navbar==0.1.5
edc-offstudy==0.2.6
edc-pdutils==0.1.3
edc-permissions==0.1.2
edc-pharmacy==0.1.2a2
edc-pharmacy-dashboard==0.1.1a3
edc-prn==0.1.5
edc-protocol==0.1.8
edc-reference==0.1.18
edc-registration==0.2.15
edc-reportable==0.1.6
edc-reports==0.1.6
edc-rest==0.1.3
edc-search==0.1.7
edc-selenium==0.1.1a1
edc-subject-dashboard==0.1.10
edc-timepoint==0.1.6
edc-visit-schedule==0.2.10
edc-visit-tracking==0.2.11
